Fix Rust types to match OpenAPI spec and actual API behavior
- Convert FeeType, FeeUnit, FeeCollection from string to integer enums
- Rewrite TeamLogin struct to match OpenAPI permissions schema
- Remove unverified ChargebackStatusValue variants (New, UnderReview,
  Responded, Expired) - only Open, Won, Lost, Closed are verified
- Add flexible deserializers for transaction enums (accept int or string)
- Make Customer.merchant optional (API returns null in some contexts)
- Add integration tests for enum value discovery via raw JSON inspection
- Document API inconsistencies in API_INCONSISTENCIES.md
- Add OpenAPI 3.1 spec for reference

Test results: 47/48 passing (TeamLogins fails due to API permission issue)

+++ src/types/disbursement_entry.rs
@@ -10,7 +10,7 @@ use super::{bool_from_int_default_false, DisbursementStatus, PayrixId};
 /// A Payrix disbursement entry.
 ///
 /// Entries track individual fund movements within a disbursement.
-/// All monetary values are in **cents**.
+/// NOTE: Monetary values may be floating point despite documentation suggesting cents.
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 #[cfg_attr(feature = "sqlx", derive(sqlx::FromRow))]
 #[serde(rename_all = "camelCase")]
@@ -38,17 +38,17 @@ pub struct DisbursementEntry {
     #[serde(default)]
     pub status: Option<DisbursementStatus>,
 
-    /// Entry amount in cents
+    /// Entry amount (may be float due to API inconsistencies)
     #[serde(default)]
-    pub amount: Option<i64>,
+    pub amount: Option<f64>,
 
-    /// Fee amount in cents
+    /// Fee amount (may be float due to API inconsistencies)
     #[serde(default)]
-    pub fee: Option<i64>,
+    pub fee: Option<f64>,
 
-    /// Net amount after fees in cents
+    /// Net amount after fees (may be float due to API inconsistencies)
     #[serde(default)]
-    pub net: Option<i64>,
+    pub net: Option<f64>,
 
     /// Currency code (e.g., "USD")
     #[serde(default)]
@@ -123,9 +123,9 @@ mod tests {
         assert_eq!(entry.merchant.unwrap().as_str(), "t1_mer_12345678901234567890123");
         assert_eq!(entry.fund.unwrap().as_str(), "t1_fnd_12345678901234567890123");
         assert_eq!(entry.status, Some(DisbursementStatus::Processed));
-        assert_eq!(entry.amount, Some(10000));
-        assert_eq!(entry.fee, Some(100));
-        assert_eq!(entry.net, Some(9900));
+        assert_eq!(entry.amount, Some(10000.0));
+        assert_eq!(entry.fee, Some(100.0));
+        assert_eq!(entry.net, Some(9900.0));
         assert_eq!(entry.currency, Some("USD".to_string()));
         assert_eq!(entry.entry_type, Some("payment".to_string()));
         assert_eq!(entry.txn.unwrap().as_str(), "t1_txn_12345678901234567890123");
